make wc eth chain id u64

+++ mm2src/coins/eth.rs
@@ -2844,7 +2844,7 @@ async fn sign_raw_eth_tx(coin: &EthCoin, args: &SignEthTransactionParams) -> Raw
                     gas: args.gas_limit,
                     data: &data,
                     nonce,
-                    chain_id: coin.chain_id as u8,
+                    chain_id: coin.chain_id,
                 })
                 .await
                 .mm_err(|err| RawTransactionError::TransactionError(err.to_string()))?;
